move main back into __init__
